==> scripts/0_Full_analysis.Rmd <==
---
title: "Run full analysis"
author: "S. Jean"
date: "2025-06-11"
output: html_document
---

```{r, include=FALSE}
knitr::opts_chunk$set(echo = F)
library(here)

```




```{r}

rm(list = ls())

rmd_files = list.files(here("scripts"), 
                       pattern = "^[^0].*\\.Rmd$",
                       full.names = T,
                       ignore.case = TRUE)
i = 1
for (rmd in rmd_files) {
  # Path to original .Rmd
  # Temporary .R path (in memory or a temp dir)
  temp_r <- tempfile(fileext = ".R")
  
  # Extract R code from .Rmd
  purl(input = rmd, output = temp_r, quiet = TRUE)
  
  # Run it
  source(temp_r, echo = TRUE, print.eval = TRUE)
  
  print(paste0(rmd_files[i], " is over"))
  i = i+1
}

```
