==> scripts/1_Parameters.Rmd <==
---
title: "Parameters"
author: "Lawson et al."
date: "2023-06-16"
output: html_document
---
```{r setup, include=FALSE}
# Setup global chunk options for knitr
knitr::opts_chunk$set(echo = FALSE)
options(scipen = 999)  # Disable scientific notation


# --- Libraries ---

# Core tidyverse packages
library(tidyverse)    # dplyr, tidyr, ggplot2, stringr, etc.

# Data presentation and tables
library(kableExtra)
library(stargazer)
library(tinytable)

# Plotting & colors
library(ggsci)
library(wesanderson)
library(RColorBrewer)
library(colorspace)
library(scales)
library(patchwork)

# Fonts and styling
library(extrafont)
font_ <- 'Helvetica'   # Set font variable for styling

# Project workflow and paths
library(here)

# Parallel processing
library(parallel)

# Web / rendering helpers
library(chromote)
library(webshot2)
library(writexl)

# Additional utility
library(berryFunctions)
library(stringi)
# Specialized domain packages
if (!("FishLife" %in% installed.packages())) {
  devtools::install_github("james-thorson/FishLife")
}
library(FishLife)

# Color palette for plots
color_growth <- 'black'
color_vertical_monopoly <- '#B22222'
color_bertrand <- '#1f78b4'
color_cournot <- '#248721'
palette_ <- c(color_growth, color_vertical_monopoly, color_bertrand, color_cournot)

# Latex flag (usage not shown here, but kept for context)
Latex <- TRUE

```

In this script, we share the code to replicate our parameters.

# 0. File architecture

Create architecture for scripts, visuals and data. Includes `sensitivity_analysis_lq` 

```{r}
if(dir.exists(paste0(getwd(),"/scripts/"))==F){
  dir.create(paste0(getwd(),'/scripts/'))
}
if(dir.exists(paste0(getwd(),'/visuals/'))==F){
  dir.create(paste0(getwd(),'/visuals/'))
  dir.create(paste0(getwd(),'/visuals/supplementary_visuals/'))
}

if(!dir.exists(here('data', 'outputs', 'sensitivity_analysis_lq'))){
  dir.create(here('data','outputs', 'sensitivity_analysis_lq'))
  print('Folder created as data/outputs/sensitivity_analysis_lq :')
}

```

# I. Load data

* `dat_p.csv` : price & quantity data from seizures (Source: collated by the Environmental Investigation Agency (EIA) and processed for analysis by authors).

* `totoaba_biomass_catch_20230516.csv` : historical biomass and catch (Source: INAPESCA 2020).

* `w_par.csv`: cost data from for poaching costs (see Supplementary Materials).

* `v_par.csv`: cost data from for aquaculture costs (see Supplementary Materials).

```{r, Load data}

# Price quantity data from seizures
dat_p = read.csv(here("data","inputs","dat_p.csv"))
dat_p = dat_p %>%
  mutate(price_buche = g*p,
         price_kg = p*1000, 
         price_buche_kg = price_buche*1000)

# Biomass and catch data
dat_b = read.csv(here('data', 'inputs', "totoaba_biomass_catch_20230516.csv")) %>% 
  rename(y = X) %>% 
  rename(b = Biomass..Tonnes.) %>% 
  rename(c = Catch..Tonnes.)

# Fishery operational costs data
w_par = read.csv(here("data", 'inputs', "w_par.csv"))

# Aquaculture operational costs data
v_par = read.csv(here("data", 'inputs', "v_par.csv")) 

sources = read.csv(here("data", 'inputs', "sources.csv"))

```


# II. Demand

```{r, Demand}
# Set the typical weight of totoaba, and typical weight of swim bladder
avg_toto_weight = 25.7
avg_buche_weight = .5

buche_to_toto = avg_buche_weight/avg_toto_weight

# Merge observed price data with observed catch and biomass data from 2013 to 2017
new_dat = merge(dat_b %>%
                  subset(y>2013) %>% 
                  select(y,
                         c)%>%
                  rename(year=y), 
                dat_p %>% 
                  select(year, 
                         price_kg))%>%
  mutate(price_ton = price_kg * 1000, 
         price_ton_toto = price_ton * buche_to_toto)

# Graphical output
new_dat %>%
  ggplot(aes(x = c, y = price_ton_toto))+
  geom_point(aes(colour = as.factor(year)), size = 3)+
  geom_smooth(method = 'lm', colour = 'black')+
  theme_bw()+
  theme(legend.position = 'bottom')+
  xlab('Catch')+
  ylab('Price per ton of totoaba in USD')+
  scale_color_aaas()+
  labs(colour = NULL)
ggsave(here('visuals','supplementary_visuals','demand_data.jpg' ))

# OLS Model for price of totoaba complete, not just buche. 
model  = lm(price_ton_toto ~ c, data = new_dat)
summary(model)

# Coefficients extraction
alpha = model$coefficients[1]
beta = abs(model$coefficients[2])

# Product differentiation at k% => gamma = sqrt(k)*beta
gamma = sqrt(.75)*beta

```


# II. Biological parameters
```{r, Biological}
# Totoaba macdonaldi is in FishLife database, so prediction for 'r' is informed by species-specific data.
Predict = Plot_taxa(Search_species(Genus="Totoaba",Species="macdonaldi")$match_taxonomy, mfrow=c(2,2))

Predict[[1]]$Mean_pred %>%
  kbl() %>%
  kable_styling()

r = Predict[[1]]$Mean_pred['r']


k <- 20226 #carrying capacity (biomass in mt) (source: INAPESCA 2020) 

# Catchability : catch per trip and stock
max_E = sum(w_par$total_fleet_days_year)

# Estimation of catchability
catchability = dat_b %>% 
  subset(y==2016) %>% 
  select(c)/(max_E * dat_b %>% 
               subset(y==2016) %>% 
               select(b))
catchability = pull(catchability)
```

# III. Poaching costs

## A. Build total cost function

### 1. Build average subseasonal costs

First : 

* Use `w_par`to reconstruct costs of crew earning, food, fuel, gear

* Compute a cost per trip, with confiscations of gear and vessels

* Compute sub-seasonal average costs


```{r calculating cost of poaching}
# W parameter, proportional increase in cost per unit (trip)
# Step one, get all annual operating costs
# Goal is to calculate cost per vessel trip (we have total fleet days per year which is number of vessels * fishing days per month)
# First calculate total annual costs for the entire fleet

w_par_ops <- w_par %>% 
  mutate(cost_crew_earnings = (totoaba_year * price_fishers_per_totoaba * monthly_earnings_paid)) %>% 
  mutate(cost_food_fuel = (foodfuel_day * total_fleet_days_year)) %>% 
  mutate(cost_gear = total_fleet_days_year*(gear_loss_day*gear_replacement)) %>% 
  mutate(revenue_per_month = (totoaba_year*(price_fishers_per_totoaba)) - (cost_crew_earnings-cost_food_fuel-cost_gear))

# Next, divide by total fishing days per year

w_par_val <- w_par_ops %>% 
  select(vessels, totoaba_day, total_fleet_days_year, cost_crew_earnings,cost_food_fuel,cost_gear,bribes_year,revenue_per_month) %>% 
  mutate(cost_per_trip = (cost_crew_earnings + cost_food_fuel + cost_gear + bribes_year)/total_fleet_days_year) 

#Adding in gear and vessel confiscation (calculated as a shared cost applied to the entire fleet)

w_pal_val_conf <- w_par_val %>%
  add_column(net_confs_cost_fleetday=664000/1920) %>% #Total cost as a proportion of reported confiscations averaged over the year (1,920 fishing trips total). 5,720 nets in the water each year (each vessel sets 2-3 nets), and 0.07 of those nets are confiscated (we know that there are 415 gear confiscation events each year). So if we assume that each net costs $1,600, then our total cost to the fleet is $664,000.
  add_column(vessel_confs_cost_fleetday=1230000/1920) %>% #Total cost as a proportion of reported vessel seizures averaged over the year (1,920 fishing trips total). There are 150 boats in total each year, and 0.546 are confiscated (there are 82 confiscations a year). So we assume each boat is worth $15,000, and therefore the total cost to the fleet is $1,230,000.
  add_column(fines_cost_fleetday=81675/1920) %>% #Total cost as a proportion of reported vessel seizures averaged over the year (1,920 fishing trips total). There are 11 arrests/convictions where people are fined $7,425/event, which would add an additional cost of $42.54.
  mutate(cost_per_trip = (cost_crew_earnings + cost_food_fuel + cost_gear + bribes_year)/total_fleet_days_year) %>% 
  mutate(cost_per_trip_wconf = cost_per_trip + net_confs_cost_fleetday + vessel_confs_cost_fleetday + fines_cost_fleetday)

costs = w_pal_val_conf %>% 
  select(cost_per_trip_wconf)%>%
  unique()%>%
  drop_na()%>%
  pull()%>%
  sort()
# Final estimates of cost per trip (W)

W_low = costs[1] #Low Season 
W_mid = costs[2] # Mid Season
W_high = costs[3] # High Season
```



### 2. Total cost from seasonal average cost


We have average cost data, for bins of trips. Here, I estimate a cost function for all the effort levels, that maps the number of trips to their total cost, taking into account different bins.

```{r, new costs }

# Find bins 
counts = c(0)
i = 2
for(x in costs){
  counts = append(counts, counts[i-1] + 
    w_pal_val_conf%>%
    subset(cost_per_trip_wconf == x)%>%
    select(total_fleet_days_year)%>%
    sum())
  i = i+1
}
counts
# Assign average cost to number of trip 
av_binned_costs = rep(1, max_E)
costs2 = data.frame(trip = seq(1,max_E), av_binned_costs)
costs2[1: 40, 'av_binned_costs'] = costs[1]
costs2[41: 520, 'av_binned_costs'] = costs[2]
costs2[521: 1920, 'av_binned_costs'] = costs[3]

# Find total cost for fleet
costs2 = costs2 %>%
  mutate(total_cost = av_binned_costs * trip,
         squared_trip = trip^2)

costs2 %>%
  ggplot(aes(x = trip, y = total_cost))+
  geom_line()


# If we had a constant marginal cost, it would take the largest value at the end
costs[3]
```
The kinky total cost function could be best approximated with a linear term. However, to model the increasing opportunity costs of having more fishermen, it would be nice to have a quadratic term, to forecast the future. 

## B. Estimate $W, W_1, W_2$

### 1. Estimate $W$

Find the parameter $W_{quad}$ such that :
$$
TotCost(E)=WE^2 \Rightarrow MargCost(E) = 2W_{quad}E
$$

And 

$$
MargCost(E_{hist}) = AvgCost(E_{hist})\Rightarrow W_{quad} = \frac{W_{avg}}{2E_{hist}}
$$

```{r, Quadratic costs}
# Given the reconstructed cost data, it is more sound to use 

W_qlow = W_low/(2*max_E)
W_qmid = W_mid/(2*max_E)
W_qhigh = W_high/(2*max_E)

# Use the reconstructed cost data and look at ways such that :
# Marginal cost at historical levels are equal
# And observed levels at historical levels are visible
costs2_future = data.frame(trip = seq((max_E+1), 5000))%>%
  mutate(squared_trip = trip^2,
         total_cost = costs[3]*trip)

costs_here = costs2 %>%
                     select(- av_binned_costs)

costs_here = rbind(costs_here, 
                   costs2_future)

# Needs to be saved
cost_differential = costs2 %>%
  mutate(quad_cost = W_qhigh*squared_trip)%>%
  mutate(differential = total_cost - quad_cost)%>%
  subset(trip == max_E)%>%
  select(differential)%>%
  pull()%>%
  unique()
    
# Graph
costs_here %>%
  mutate(quad_cost = W_qhigh*squared_trip + cost_differential)%>%
  pivot_longer(- trip, 
               names_to = 'names', 
               values_to = 'values')%>%
  subset(!(names %in% c('squared_trip')))%>%
  ggplot(aes(x = trip, y = values, color = names))+
  geom_line()+
  geom_vline(xintercept = max_E, linetype = 'dotted')+
  scale_color_manual(name = 'Cost', labels = c('Quadratic cost', 'Linear cost'), values = c('royalblue', 'black') )+
  theme_bw()
```


### 2. Estimate $W_1, W_2$



First, find best estimate of modeled cost in a MMSE sense. 

```{r, new_costs2}
# We must have that the marginal cost at the end with a linear quadratic model is W1 + 2 W2 E_hist = AVG
potential_W1 = seq(0, costs[3]+1)
potential_W2 = (costs[3] - potential_W1)/2

least_squares = data.frame(0,0)
colnames(least_squares) = c('value', 'index')

try_ls = F
if(try_ls == T){
  for(row in 1:length(potential_W1)){
  costs2 = costs2 %>%
    mutate(attempt = potential_W1[row] * trip + potential_W2[row]*squared_trip)
  least_square = costs2 %>%
    mutate(ls = (total_cost - attempt)^2)%>%
    select(ls)%>%
    sum()
  least_squares = rbind(least_squares, c(least_square, row))
  }
  write.csv(least_squares, here('data','outputs', 'least_squares_W.csv'))
}else{
  least_squares = read.csv(here('data', 'outputs', 'least_squares_W.csv'))
}

#least_squares = least_squares[-1, ]
#print(least_squares[which(least_squares$value == min(least_squares$value)),])
print(paste0('Optimal W1 in an MMSE sense is : ', potential_W1[which(least_squares$value == min(least_squares$value))]))
print(paste0('Optimal W2 in an MMSE sense is : ', potential_W2[which(least_squares$value == min(least_squares$value))]))
```

Result from MMSE suggest to use the linear form, without reference to a quadratic term. However, increasing cost of effort is consistent with a disaggregated view of the fishery. Force a quadratic term.


```{r, other method}

# Set baseline parameter choice. 

# Set potential costs
W1_potential = seq(11000,W_high, by = 400)
W2_potential = (W_high - W1_potential)/(2*max_E)

store = data.frame(0,0,0,0)
colnames(store) = c('W1',"W2","diff%","mse")

for(baseline_W_choice in 1:length(W1_potential)){
  W_lq_new = data.frame(W1 = W1_potential, W2 = W2_potential)%>%
    mutate(AC = costs[3], 
           W_mid = W_high/(2*max_E), 
           baseline_choice = baseline_W_choice)
  
  # Compute historical costs, and projected costs, and projected linear quadratic costs. 
  costs2_historical  = costs2 %>%
    mutate(estimated_linear_quadratic = W1_potential[baseline_W_choice] * trip + W2_potential[baseline_W_choice]* squared_trip)%>%
    select(trip, estimated_linear_quadratic, total_cost)
  
  costs2_future = data.frame(trip = seq(max_E, 5000))%>%
    mutate(squared_trip = trip^2,
           total_cost = costs[3]*trip, 
           estimated_linear_quadratic = W1_potential[baseline_W_choice] * trip + W2_potential[baseline_W_choice] * squared_trip)%>%
    select(trip, estimated_linear_quadratic, total_cost)
  
  costs_final = rbind(costs2_historical, costs2_future)
  rm(costs2_historical, costs2_future)
  # Plot costs and estimate. 
  #costs_final%>%
  #  pivot_longer(-trip, 
   #           values_to = 'values', 
 #                names_to =  'Costs')%>%
  #  ggplot(aes(x = trip, y= values, colour = Costs))+
   # geom_line()+
  #  geom_vline(xintercept = max_E, linetype = 'dotted')+
  #  scale_color_manual(name = 'Cost', labels = c('Linear quadratic cost', 'Linear cost'), values = c('royalblue', 'black') )+
  #  theme_bw()
  
  costs_final= costs_final %>%
    mutate(diff = total_cost - estimated_linear_quadratic, 
           diff2 = diff^2,
           diff_pct = diff/total_cost*100)
  diff= costs_final%>%
    subset(trip == 1920)%>%
    select(diff_pct)%>%
    pull()%>%
    unique()
  mse = costs_final%>%select(diff2)%>%sum()
  dat = c(W1_potential[baseline_W_choice], W2_potential[baseline_W_choice], diff, mse)
  store = rbind(store, dat)
}

kable(store)

```

Here, I keep as the baseline specification the first value with low difference with cost

```{r, baseline choice for W}
baseline_W_choice = 4

W_lq_new = data.frame(W1 = W1_potential, W2 = W2_potential)%>%
    mutate(AC = costs[3], 
           W_mid = W_high/(2*max_E), 
           baseline_choice = baseline_W_choice)
  
  # Compute historical costs, and projected costs, and projected linear quadratic costs. 
  costs2_historical  = costs2 %>%
    mutate(estimated_linear_quadratic = W1_potential[baseline_W_choice] * trip + W2_potential[baseline_W_choice]* squared_trip)%>%
    select(trip, estimated_linear_quadratic, total_cost)
  
  costs2_future = data.frame(trip = seq(max_E, 5000))%>%
    mutate(squared_trip = trip^2,
           total_cost = costs[3]*trip, 
           estimated_linear_quadratic = W1_potential[baseline_W_choice] * trip + W2_potential[baseline_W_choice] * squared_trip)%>%
    select(trip, estimated_linear_quadratic, total_cost)
  
  costs_final = rbind(costs2_historical, costs2_future)
  rm(costs2_historical, costs2_future)
  # Plot costs and estimate. 
  costs_final%>%
    pivot_longer(-trip, 
                 values_to = 'values', 
                 names_to =  'Costs')%>%
    ggplot(aes(x = trip, y= values, colour = Costs))+
    geom_line()+
    geom_vline(xintercept = max_E, linetype = 'dotted')+
    scale_color_manual(name = 'Cost', labels = c('Linear quadratic cost', 'Linear cost'), values = c('royalblue', 'black') )+
    theme_bw()

```

### 3. Export data and sources for W

Export data for $W$ and the corresponding model parameters for the quadratic model ($W$) and linear-quadratic model ($W_1$ and $W_2$)

```{r supplementary materials tables}
classer = function(x){
  if(x %in% c(1,6)){
    return('Low')
  }else if(x %in% c(2,5)){
    return('Mid')
  }else if(x %in% c(3,4)){
    return('High')
  }
}
#w_par_b = w_pal_val_conf %>% 
w_par_b = w_par %>% 
  mutate(month = seq(1:nrow(w_pal_val_conf)))%>%
    subset(vessels>0)%>%
  mutate(season = sapply(month, classer))%>%
  group_by(season)%>%
  summarise(across(everything(),mean),
            .groups= 'drop')%>%
  t()%>%
  as.data.frame()

w_par_b = w_par_b%>%
  mutate(name = rownames(w_par_b))

sources_w = sources %>% 
  subset(parameter=='W') %>% 
  select(-parameter)

w_par_b = merge(w_par_b, sources_w, by = 'name', all=T)

colnames(w_par_b) = c('Variable', 'High', 'Low', 'Mid', 'Source')

order__ = c('vessels', 'days permonth', 'total fleet days year', 'foodfuel day', 'totoaba gearset', 'gear loss day', 'gearset vessel perday', 'gear replacement', 'bribes year', 'Average cost')

w_par_b = w_par_b %>% 
  drop_na()%>%
  rbind(append('Average cost', append(round(costs,2), 'Own comp' )))%>%
  subset(!(Variable %in% c('crew_size','month', 'monthly_earnings_paid', 'price_fishers_per_totoaba',
                           'totoaba_year')))%>%
  mutate(Variable = str_replace_all(Variable, '_', ' '), 
         Source = replace(Source, Source == 'INAPESCA 2020', 'cite{INAPESCA}'))%>%
  arrange(match(Variable, order__))%>%
  relocate(Variable, Low, Mid, High, Source)


if(Latex==T){
  w_par_b %>%
    kable(caption ='Information supporting $W$' ,
          format = 'latex', 
          align = 'c')
}

write.csv(W_lq_new, here('data', 'inputs', 'W_lq_new.csv'), row.names = F)

```



# IV. Trading costs

```{r trading costs}
c = 0
```
We use `c=` `r c` as trading costs.

# V. Farming costs

```{r calculating cost of farming}
# Farming cost data, from Buche Bros

# Final estimate per tonne of totoaba per year:
v_ = v_par %>% 
  select(op_cost_per_t)%>%
  pull()

age = 4.5
v = v_*age

# Capitalized costs with interest rate
interest_rate = 0.1

v_cap = v_*(1 + (1 + interest_rate) + (1 + interest_rate)^2 + (1 + interest_rate)^3 + (1+ interest_rate)^3.5)
v = v_cap
# Calculating size at harvest
aqua_par_w <- read.csv(here("data", 'inputs', "aqua_par_wild.csv"))
aqua_par_f <- read.csv(here("data", 'inputs', "aqua_par_farm.csv"))
```
At 4.5 years under farming conditions, a totoaba should have a `r aqua_par_f %>% subset(Totoaba_Age==4.5) %>% select(Weight_Buche_g) %>% pull() %>% round(2)`g swim bladder. At 5 years under farming conditions, a totoaba should have a `r aqua_par_f %>% subset(Totoaba_Age==5) %>% select(Weight_Buche_g) %>% pull() %>% round(2)` g swim bladder

```{r table for v}

v_par_b = v_par %>% 
  t() %>% 
  as.data.frame()
v_par_b = v_par_b %>%
  mutate(name = rownames(v_par_b))

order__ = rownames(v_par_b)

sources_v = sources %>% 
  subset(parameter=='v') %>% 
  select(-parameter)

v_par_b = merge(v_par_b, sources_v, by='name',  all = T)%>%
  arrange(match(name, order__))%>%
  mutate(name = str_replace_all(name, '_', ' '))
if(Latex==T){
  v_par_b %>%
    kable(caption ='Information supporting $v$' ,
          format = 'latex', 
          align = 'c')
}
```


## Quasi-fixed cost for set-up

The quasi fixed cost for pens is `cost_pen` from the Buche Bros data

```{r}

cost_pen = 200000

```

# VI. Save calibration data
```{r}


params = c('alpha', 
           'beta',
           'r',
           'k',
           'sigma',
           'avg_cost',
           #'W_mid',
           #'W_low',
           'W_high', 
           'fixed_cost',
           'W1', 
           'W2', 
           'age', 
           'gamma', 
           'v', 
           'i_r',
           'c',
           'cost_pen')
value = c(alpha,
          beta, 
          r, 
          k, 
          catchability,
          W_high,
          #W_qmid,
          #W_qlow,
          W_qhigh, 
          cost_differential,
          W_lq_new[baseline_W_choice,'W1'],
          W_lq_new[baseline_W_choice, 'W2'], 
          age, 
          gamma,
          v, 
          interest_rate, 
          c,
          cost_pen)
concept = c('Demand model : intercept', 
            'Demand model : coefficient', 
            'Intrinsic growth rate', 
            'Carrying capacity (in metric tons)',
            'Catchability', 
            'Average cost per vessel trip at historical value', 
            #'Parameter cost (MC = Avg) at historical value - middle season', 
            #'Parameter cost (MC = Avg) at historical value - low season', 
            'Quadratic cost parameter - Quadratic cost function', 
            'Fixed cost - Quadratic cost function',
            'Linear cost parameter - Linear quadratic cost function', 
            'Quadratic cost parameter - Linear quadratic cost function',
            'Age of farmed totoaba', 
            'Demand model : substitutable good coefficient', 
            'Unit cost of farming', 
            'Interest rate',
            'Unit cost of trading',
            "Cost of pen")

units = c('USD', 
          'USD/metric ton of biomass',
          'unitless',  
          'metric tons of biomass',
          '% of biomass/vessel trip',
          'USD/vessel trip', 
          #'USD vessel trip$^{-2}$',  
          #'USD vessel trip$^{-2}$', 
          'USD vessel trip$^{-2}$', 
          'USD',
          'USD/vessel trip', 
          'USD vessel trip$^{-2}$', 
          'Years', 
          'USD/metric ton of biomass', 
          'USD/metric ton of biomass', 
          '%', 
          'USD/ metric ton of biomass',
          "USD")

calibration = data.frame(params, value, concept, units)
write.csv(calibration, here('data', 'outputs', 'calibration_params.csv'))
```

```{r, calibration2}
calibration %>%  
  mutate(value = round(value, 2))%>%
  kbl() %>%
  kable_styling()

Latex = T

if(Latex==T){
  calibration %>%
    mutate(value = round(value, 2))%>%
    kable(caption = 'Parameters for calibration', 
          format = 'latex', 
          align = 'c')%>%
    kable_minimal(full_width = F)
}
```
