==> scripts/4_Additional_graphs.Rmd <==
---
title: "Additional graphs and analysis"
author: "Lawson et al."
date: "2023-06-19"
output: html_document
---

```{r, include=FALSE}
knitr::opts_chunk$set(echo = F)
options(scipen=999)

#result1 = read.csv(here('data', 'outputs', 'result1.csv'))


calibration = read.csv(here("data", 'outputs', "calibration_params.csv"))

for(i in 1:nrow(calibration)){
  assign(calibration$params[i], calibration$value[i])
}  


# Build the 2nd order parameters from parameter list
  # Differentiate parameters for Wild and Farmed inverse demands
alpha_w = alpha
alpha_f = alpha
beta_w = beta
beta_f = beta
  # Demand functions parameters
e   = gamma/(beta_w*beta_f - (gamma^2))
a_f = (alpha_f*beta_w - alpha_w*gamma)/(beta_w*beta_f - (gamma^2))
a_w = (alpha_w*beta_f - alpha_f*gamma)/(beta_w*beta_f - (gamma^2))
b_f = beta_f/(beta_w*beta_f - (gamma^2))
b_w = beta_w/(beta_w*beta_f - (gamma^2))

if(dir.exists(here('visuals/supplementary_visuals'))==F){
  dir.create(here('visuals/supplementary_visuals/'))
}

```

## Figure (SM):
```{r, Fig1}
dat_b = read.csv(here('data', 'inputs', "totoaba_biomass_catch_20230516.csv")) %>% 
  rename(y = X) %>% 
  rename(b = Biomass..Tonnes.) %>% 
  rename(c = Catch..Tonnes.)

scale = 10
dat_b %>% 
  ggplot(aes(x=y))+
  geom_line(aes(y = c, color = 'Catch (metric tons)'), linewidth = 1.1)+
  geom_line(aes(y = b/scale, color = 'Biomass (metric tons)'), linewidth = 1.1)+
  scale_y_continuous(sec.axis = sec_axis(~.*scale, name='Biomass (metric tons)'))+
  geom_vline(xintercept = 1976, color = 'grey', linewidth = 1, linetype ='dashed')+
  geom_vline(xintercept = 2010, color = 'grey', linewidth = 1, linetype ='dashed')+
  scale_color_npg()+
  theme_bw()+
  ylab('Catch in tons of biomass')+
  xlab('year')+
  theme(legend.position = 'bottom',
        axis.text.x = element_text(angle = 70, 
                                   hjust = 1, 
                                   family = font_),
        axis.title.x = element_text(family = font_),
        axis.text.y = element_text(family = font_),
        axis.title.y = element_text(family = font_), 
        strip.text = element_text(family = font_),
        legend.text = element_text(family = font_)
        )+
  labs(color=NULL)
ggsave(here("visuals","supplementary_visuals", "Trends_stock_catch.png"))
```

## Figure (SM) and 1 in Main : Hypothetical schematic of equilibrium points under different poaching harvest functions
```{r}
qty_monop = function(x){
 y = (400 - 4)*.02^2 * x^2 / (2*80* 0.02^2 * x^2 + 2*5)
 return(y)
}

qty_monop2 = function(x){
 y = (400 - 4)*.02^2 * x^2 / (2*80* 0.025^2 * x^2 + 2*5)
 return(y)
}

qty_monop3 = function(x){
 y = (400 - 4)*.02^2 * x^2 / (2*80* 0.018^2 * x^2 + 2*5)
 return(y)
}

qty_monop4 = function(x){
 y = (10 - 4)*.25^2 * x^2 / (2*80* 0.02^2 * x^2 + 2*5)
 return(y)
}

growth = function(x){
  y = 0.1*x*(1-x/100)
  return(y)
}

size_ = 5

points_ = data.frame(x = c(7.9, 30.46, 61.15, 80.75),
                     y = c(0.75, 2.12 , 2.36, 1.56))

data.frame(x = seq(1,100))%>%
  mutate(monop = qty_monop(x),
         monop2 = qty_monop2(x),
         monop3 = qty_monop3(x),
         monop4 = qty_monop4(x),
         growth = growth(x))%>%
  ggplot(aes(x = x))+
  geom_line(aes(y = monop, colour = 'Harvest'), linewidth = 1.2)+
  geom_line(aes(y = monop2, colour = 'Harvest'), linewidth = 1.2)+
  geom_line(aes(y = monop3, colour = 'Harvest'), linewidth = 1.2)+
  #geom_line(aes(y = monop4, colour = "Harvest"), linewidth = 1.2)+
  geom_line(aes(y = growth, colour = 'Growth'), linewidth = 1.2)+
  geom_point(data = points_, aes(x=x, y=y), size=3, shape = 17)+
  scale_color_manual(values = c( 'grey','black'))+
  #ylim(0,3)+
  ylab('Growth / harvest')+
  xlab('Stock')+
  theme_classic()+
  #geom_segment(x = 8, xend = 8, y = 0, yend= .75, linetype = 2)+
  #geom_segment(x = 31, xend = 31, y = 0, yend= 2.1, linetype = 2)+
  #geom_segment(x = 61, xend = 61, y = 0, yend= 2.35, linetype = 2)+
  annotate('text', x = 15, y = .7, label= 'stable', family = font_, size = size_)+
  annotate('text', x = 38, y = 2, label= 'unstable', family = font_, size = size_)+
  annotate('text', x = 65, y = 2.5, label= 'stable', family = font_, size = size_)+
  annotate('text', x = 85, y = 1.7, label= 'stable', family = font_, size = size_) +
  annotate('text', x = 103, y = 3, label= 'A', family = font_, size = size_) +
  annotate('text', x = 103, y = 2.45, label= 'B', family = font_, size = size_) +
  annotate('text', x = 103, y = 1.55, label= 'C', family = font_, size = size_) +
  theme(legend.position = "none", 
        axis.ticks.x = element_blank(),
        axis.ticks.y = element_blank(),
        axis.text.x = element_blank(),
        axis.text.y = element_blank(),
        strip.text.x  = element_text(family = font_) ,
        axis.title.x  = element_text(family = font_),
        axis.text = element_text(family = font_),
        axis.title.y  = element_text(family = font_),
        legend.text = element_text(family = font_), 
        legend.title = element_text(family = font_))

ggsave(here("visuals", "supplementary_visuals", "Figure1_potential_equilibria.jpg"), 
       width = 15, height = 30, units = 'cm')
```
## Figure (SM) and 1 in Main : Alternative with extinction - Hypothetical schematic of equilibrium points under different poaching harvest functions

```{r}
qty_monop = function(x){
 y = (400 - 4)*.02^2 * x^2 / (2*80* 0.02^2 * x^2 + 2*5)
 return(y)
}

qty_monop2 = function(x){
 y = (400 - 4)*.02^2 * x^2 / (2*80* 0.025^2 * x^2 + 2*5)
 return(y)
}

qty_monop3 = function(x){
 y = (400 - 4)*.02^2 * x^2 / (2*80* 0.018^2 * x^2 + 2*5)
 return(y)
}

qty_monop4 = function(x){
  y = (10 - 4) * 0.5^2 * x^2 / (2 * 500 * 0.02^2 * x^2 + 2 * 5)
 return(y)
}

growth = function(x){
  y = 0.1*x*(1-x/100)
  return(y)
}

size_ = 5

points_ = data.frame(x = c(0,7.9, 30.46, 61.15, 80.75),
                     y = c(0,0.75, 2.12 , 2.36, 1.56))

data.frame(x = seq(0,100))%>%
  mutate(monop = qty_monop(x),
         monop2 = qty_monop2(x),
         monop3 = qty_monop3(x),
         monop4 = qty_monop4(x),
         growth = growth(x))%>%
  ggplot(aes(x = x))+
  geom_line(aes(y = monop, colour = 'Harvest'), linewidth = 1.2)+
  geom_line(aes(y = monop2, colour = 'Harvest'), linewidth = 1.2)+
  geom_line(aes(y = monop3, colour = 'Harvest'), linewidth = 1.2)+
  geom_line(aes(y = monop4, colour = "Harvest"), linewidth = 1.2)+
  geom_line(aes(y = growth, colour = 'Growth'), linewidth = 1.2)+
  geom_point(data = points_, aes(x=x, y=y), size=3, shape = 17)+
  scale_color_manual(values = c( 'grey','black'))+
  #ylim(0,3)+
  ylab('Growth / harvest')+
  xlab('Stock')+
  theme_classic()+
  #geom_segment(x = 8, xend = 8, y = 0, yend= .75, linetype = 2)+
  #geom_segment(x = 31, xend = 31, y = 0, yend= 2.1, linetype = 2)+
  #geom_segment(x = 61, xend = 61, y = 0, yend= 2.35, linetype = 2)+

  theme(legend.position = "none", 
        axis.ticks.x = element_blank(),
        axis.ticks.y = element_blank(),
        axis.text.x = element_blank(),
        axis.text.y = element_blank(),
        strip.text.x  = element_text(family = font_) ,
        axis.title.x  = element_text(family = font_),
        axis.text = element_text(family = font_),
        axis.title.y  = element_text(family = font_),
        legend.text = element_text(family = font_), 
        legend.title = element_text(family = font_))+
    annotate('text', x = 15, y = .7, label= 'stable', family = font_, size = size_)+
  annotate('text', x = 38, y = 2, label= 'unstable', family = font_, size = size_)+
  annotate('text', x = 65, y = 2.5, label= 'stable', family = font_, size = size_)+
  annotate('text', x = 85, y = 1.7, label= 'stable', family = font_, size = size_) +
  annotate('text', x = 0,  y = -.1, label= 'stable', family = font_, size = size_) +
  annotate('text', x = 103, y = 3.75, label= 'A', family = font_, size = size_) +
  annotate('text', x = 103, y = 3, label= 'B', family = font_, size = size_) +
  annotate('text', x = 103, y = 2.45, label= 'C', family = font_, size = size_) +
  annotate('text', x = 103, y = 1.55, label= 'D', family = font_, size = size_)

ggsave(here("visuals", "supplementary_visuals", "Figure1_potential_equilibria_alternative.jpg"), 
       width = 20, height = 20, units = 'cm')
```


## Figure : Different equilibria
```{r, fig3}
zero = function(x){
  if(x<0){
    return(NA)
  }else{
    return(x)
  }
}

growth = function(x, 
                  k.=k, 
                  r.=r){
  y = r.*x*(1-x/k.)
  return(y)
}

#plot_ = result1 %>%
  #subset(x<500)%>%
#  ggplot(aes(x=x))+
#  geom_line(aes(y = growth), linewidth = 1)+
#  geom_line(aes(y = monop_harvest, colour = 'Vertical monopoly'), linewidth = 1)+
#  geom_line(aes(y = cournot_harvest, colour = 'Quantity adjustment (Cournot)'), linewidth = 1)+
#  geom_line(aes(y = bertrand_harvest, colour = 'Price setting (Bertrand)'), linewidth = 1)+
#  geom_line(aes(y = pt_harvest, colour = 'Multiple traders'), linewidth = 1) +
#  theme_bw()+
#  ylab('Growth / harvest')+
#  xlab('Population stock')+
#  scale_color_manual(values = c('royalblue1','lightgreen','forestgreen','firebrick'))+
#  labs(colour = ' ')+
#  theme(legend.position = 'bottom',
#        strip.text.x  = element_text(family = font_) ,
#        axis.title.x  = element_text(family = font_),
#        axis.text = element_text(family = font_),
#        axis.title.y  = element_text(family = font_),
 #       legend.text = element_text(family = font_), 
#        legend.title = element_text(family = font_))
#plot_
#ggsave(here('visuals', 'supplementary_visuals', 'Figure3.jpg'))


#plot_ = result1 %>%
#  mutate(growth2 = growth(x,r.=4.03*r, k.=5000),
#         growth3 = growth(x, r. = 10.1*r, k.=2000))%>%
#  mutate(growth2 = sapply(growth2, zero), 
#         growth3 = sapply(growth3, zero))%>%
#  subset(x<7500)%>%
#  ggplot(aes(x=x))+
  #geom_line(aes(y = growth, colour = 'Growth'), linewidth = 1)+
#  geom_line(aes(y = growth2), linewidth = 1)+
  #geom_line(aes(y = growth3, colour = 'Growth'), linewidth = 1)+
#  geom_line(aes(y = monop_harvest, colour = 'Vertical monopoly'), linewidth = 1)+
#  geom_line(aes(y = cournot_harvest, colour = 'Quantity adjustment (Cournot)'), linewidth = 1)+
#  geom_line(aes(y = bertrand_harvest, colour = 'Price setting (Bertrand)'), linewidth = 1)+
#  geom_line(aes(y = pt_harvest, colour = 'Multiple traders'), linewidth = 1) +
#  theme_bw()+
#  ylab('Growth / harvest')+
#  xlab('Population stock')+
#  scale_color_manual(values = c('royalblue1','lightgreen','forestgreen','firebrick'))+
#  labs(colour = ' ')+
#  theme(legend.position = 'bottom',
#        strip.text.x  = element_text(family = font_) ,
#        axis.title.x  = element_text(family = font_),
#        axis.text = element_text(family = font_),
#        axis.title.y  = element_text(family = font_),
#        legend.text = element_text(family = font_), 
#        legend.title = element_text(family = font_))
#plot_
#ggsave(here('visuals', 'supplementary_visuals', 'Figure3b.jpg'))

#plot_ = result1 %>%
#  mutate(growth2 = growth(x,r.=4.03*r, k.=5000),
#         growth3 = growth(x, r. = 10.1*r, k.=2000))%>%
#  mutate(growth2 = sapply(growth2, zero), 
#         growth3 = sapply(growth3, zero))%>%
#  subset(x<3000)%>%
#  ggplot(aes(x=x))+
  #geom_line(aes(y = growth, colour = 'Growth'), linewidth = 1)+
  #geom_line(aes(y = growth2, colour = 'Growth'), linewidth = 1)+
#  geom_line(aes(y = growth3), linewidth = 1)+
#  geom_line(aes(y = monop_harvest, colour = 'Vertical monopoly'), linewidth = 1)+
#  geom_line(aes(y = cournot_harvest, colour = 'Quantity adjustment (Cournot)'), linewidth = 1)+
#  geom_line(aes(y = bertrand_harvest, colour = 'Price setting (Bertrand)'), linewidth = 1)+
#  geom_line(aes(y = pt_harvest, colour = 'Multiple traders'), linewidth = 1) +
 # theme_bw()+
#  ylab('Growth / harvest')+
#  xlab('Population stock')+
#  scale_color_manual(values = c('royalblue1','lightgreen','forestgreen','firebrick'))+
#  labs(colour = ' ')+
#  theme(legend.position = 'bottom',
#        strip.text.x  = element_text(family = font_) ,
#        axis.title.x  = element_text(family = font_),
#        axis.text = element_text(family = font_),
#        axis.title.y  = element_text(family = font_),
#        legend.text = element_text(family = font_), 
#        legend.title = element_text(family = font_))

#plot_
#ggsave(here('visuals', 'supplementary_visuals', 'Figure3c.jpg'))


```

## Figure : Bertrand v. Monopoly following Damania and Bulte
```{r, fig 4}

# Define functions
a_m = alpha/beta
b_m = 1/beta
W = W_high


harvest_monop_direct = function(x, 
                                a_m. = a_m, 
                                b_m. = b_m, 
                                c.  = c, 
                                W. = W, 
                                sigma. = sigma){
  y = (a_m. - b_m. *c.)*(sigma.^2)*(x^2)/(2*sigma.^2*x^2 + 2*b_m.*W.)
  return(y)
}


q_b_paper = function(x, 
                     a_f. = a_f,
                     a_w. = a_w,
                     b_f. = b_f,
                     b_w. = b_w,
                     c. = c,
                     e. = e,
                     v. = v,
                     W. = W,
                     sigma. = sigma){
  y = ((sigma.^2)*(x^2)*b_w.*(b_f.*(2*a_w.+e.*v. - 2*b_w.*c.) + e.*(c.+a_f.)))/((sigma.^2)*(x^2)*(4*b_f.*b_w. - (e.^2)) + W.*b_w.*(2*b_f.*b_w. - (e.^2)))
  return(y)
}

bertrand_harvest = function(x,
                              sigma. = sigma, 
                              a_f. = a_f, 
                              a_w. = a_w, 
                              b_f. = b_f, 
                              b_w. = b_w, 
                              e. = e, 
                              v. = v, 
                              c. = c, 
                              W. = W_high){
  y = ((sigma.^2)*(x^2)*b_w.*(b_f.*(2*a_w.+e.*v.) + c.*((e.^2)- 2*b_f.*b_w.) + e.*a_f.))/((sigma.^2)*(x^2)*(4*b_w.*b_f. - (e.^2)) + 2*W.*b_w.*(2*b_f.*b_w. - (e.^2)))
  return(y)
}



results = data.frame(x = seq(1,k))%>%
  mutate('Bertrand wild harvest' = q_b_paper(x),
         'Vertical monopoly' = harvest_monop_direct(x, a_m. = a_m,
                                                    b_m. = b_m))%>%
  pivot_longer(cols = -c('x'),
               names_to = 'names',
               values_to = 'values')

results %>% 
  ggplot(aes(x = x, 
             y = values, 
             color = names))+
  geom_line()+
  theme_bw()+
  scale_color_manual(values = c('royalblue1','forestgreen'))+
  ylab('Harvest')+
  xlab('Population')+
  guides(
    color = guide_legend(
      title = " ",  # Title for the first column
      ncol = 1              # Number of columns
    ))+
  theme(axis.text.x = element_text(angle = 70, 
                                   hjust = 1, 
                                   family = font_),
        axis.title.x = element_text(family = font_),
        axis.text.y = element_text(family = font_),
        axis.title.y = element_text(family = font_), 
        strip.text = element_text(family = font_),
        legend.position = 'bottom',
        legend.text = element_text(family = font_)
        )
#ggsave(here('visuals', 'supplementary_visuals', 'Figure4_bertrand_monop_DB.jpg'))

```

## Figure : Comparison between monopoly and Bertrand behavior on monopolistic and duopolistic markets following Damania and Bulte
```{r, fig 5}
lemma = data.frame(x = seq(1,k))%>% # Growth
                   mutate("Vertical monopoly on a duopolistic market" = harvest_monop_direct(x, 
                                                                                    a_m. = a_w,
                                                                                    b_m = b_w), # Monopoly using the duopolistic market assumption
                   "Vertical monopoly on a monopolistic market" = harvest_monop_direct(x), # Monopoly using the monopolistic market assumption
                   "Price setting (Bertrand) on a duopolistic market" = bertrand_harvest(x),
                   'Price setting (Bertrand) on a duopolistic market and large c' = bertrand_harvest(x, c.= 100000)) %>% # Bertrand harvest using the article's formula
  pivot_longer(cols = -c('x'),
               values_to = 'values',
               names_to = 'names')


colors = c('royalblue1','royalblue4','chartreuse','forestgreen')
# Plot

lemma %>% 
  ggplot(aes(x=x, y = values, color = names))+
  geom_line()+
  scale_color_manual(values = colors, 
                     labels = c("Price setting (Betrand) a duopolistic market",
                                'Price setting (Bertrand) on a duopolistic market and large c',
                                "Vertical monopoly on a duopolistic market",
                                "Vertical monopoly on a monopolistic market"
                                ))+
  theme_bw()+
  xlab('Population ')+
  ylab("Harvest")+
  guides(
    color = guide_legend(
      title = " ",  # Title for the first column
      ncol = 2              # Number of columns
    ))+
  theme(axis.text.x = element_text(angle = 70, 
                                   hjust = 1, 
                                   family = font_),
        axis.title.x = element_text(family = font_),
        axis.text.y = element_text(family = font_),
        axis.title.y = element_text(family = font_), 
        strip.text = element_text(family = font_),
        legend.position = 'bottom',
        legend.text = element_text(family = font_)
        )
#ggsave(here('visuals', 'supplementary_visuals', 'Figure5_monop_duopoly.jpg'))

```

## Figure : F (x) and v summarise when Bertrand harvest is larger or lower than in monopoly
```{r}
for(i in 1:nrow(calibration)){
  assign(calibration$params[i], calibration$value[i])
}  

a_m = alpha/beta
b_m = 1/beta

v_func = function(x,c){
  coeff = (a_m - b_m*c)/(b_w*b_f*e)
  parenthesis = (2*W_high*b_w*(2*b_w*b_f - e^2) + (4*b_w*b_f - e^2)*sigma^2*x^2)/(2*W_high*b_m + 2*sigma^2*x^2)
  const = (e*a_f + c*(e^2 - 2*b_w*b_f) + 2*b_f*a_w)/(b_f*e)
  return(coeff* parenthesis - const)
}

tester = data.frame(x = seq(1,k))%>%
  mutate(v_func = v_func(x,c),
         v_func2 = v_func(x, 5*c),
         v = v, 
         v_ = 3.5*v,
         v_max = .1*v)



# Zoomed version
tester %>%
  subset( x > 5000)%>%
  ggplot(aes(x=x))+
  geom_line(aes(y = v_func, colour = 'F(Population)'), linewidth = 1.1)+
  geom_line(aes(y = v, colour = 'v' ), linewidth = 1.1)+
  geom_line(aes(y = v_), linewidth = 1.1, linetype = 'dotted', colour = 'red')+
  geom_line(aes(y = v_max), linewidth = 1.1, linetype = 'dotted', colour = 'lightgreen')+
  #geom_ribbon(data=subset(tester, x>=10000 & x <= 20000), 
  #            aes(ymin=v_max,ymax=1.8*v), fill="red", alpha=0.5) +
  geom_ribbon(data=subset(tester,  x>5000 & x <= 11300), 
              aes(ymin=v,ymax=v_func), fill="lightgreen", alpha=0.5) +
  geom_ribbon(data=subset(tester, x>=11300), 
              aes(ymin=v_func,ymax=v), fill="red", alpha=0.5) +
  
  theme_bw()+
  scale_color_manual(values = c('lightgreen', 'red'))+
  ylab('F(Population)')+
  xlab('Population')+
  theme(legend.position = 'bottom')+
  labs(color=NULL)+
  theme(axis.text.x = element_text(angle = 70, 
                                   hjust = 1, 
                                   family = font_),
        axis.title.x = element_text(family = font_),
        axis.text.y = element_text(family = font_),
        axis.title.y = element_text(family = font_), 
        strip.text = element_text(family = font_),
        legend.position = 'bottom',
        legend.text = element_text(family = font_)
        )
  #ggsave(here("visuals/supplementary_visuals/", "Figure6_Lemma4_zoomed.jpg"))

```





## Figure : v for poaching to stop in a Bertrand equilibrium
```{r, fig7}

gamma = seq(0.01,.99,0.01)*beta 
c = seq(0, 5000, 20)

tester = expand.grid(gamma, c)
store = data.frame()
for(row in seq(1,nrow(tester))){
  alpha_w = alpha
  alpha_f = alpha
  beta_w = beta
  beta_f = beta
  # Demand functions parameters
  gamma = tester[row, 1]
  c     = tester[row, 2]
  
  e   = gamma/(beta_w*beta_f - (gamma^2))
  a_f = (alpha_f*beta_w - alpha_w*gamma)/(beta_w*beta_f - (gamma^2))
  a_w = (alpha_w*beta_f - alpha_f*gamma)/(beta_w*beta_f - (gamma^2))
  b_f = beta_f/(beta_w*beta_f - (gamma^2))
  b_w = beta_w/(beta_w*beta_f - (gamma^2))
  
  v = 1/(e*b_f)*(c*(2*b_w*b_f-e^2) - e*a_f - 2*a_w*b_f)
  
  store = rbind(store, c(tester[row,1], tester[row,2], v))
}
colnames(store) = c( 'gamma','c', 'v')
store %>%
  subset(gamma>500)%>%
  ggplot(aes(x = c, y = gamma, fill = v))+
  geom_tile()+
    scale_fill_gradient2(low = "darkblue", high = "white", mid='lightblue')+
  theme_bw()+
  theme(axis.text.x = element_text(angle = 70, 
                                   hjust = 1, 
                                   family = font_),
        axis.title.x = element_text(family = font_),
        axis.text.y = element_text(family = font_),
        axis.title.y = element_text(family = font_), 
        strip.text = element_text(family = font_),
        legend.position = 'bottom',
        legend.text = element_text(family = font_),
        legend.title = element_text(family = font_),
        legend.key.width= unit(2, 'cm')
        )+
  labs(fill = 'V such that poaching ceases')
#ggsave(here('visuals/supplementary_visuals', 'Figure7_v_poaching_stop.jpg'))
```

## Figure : Price per ton of totoaba and catch from 2014 to 2017
```{r, fig 8}
dat_p = read.csv(here("data","inputs","dat_p.csv")) #Price-quantity data (OG Paper)
dat_p = dat_p %>%
  mutate(price_buche = g*p,
         price_kg = p*1000, 
         price_buche_kg = price_buche*1000,
         price_ton = price_kg*1000)

new_dat = merge(dat_b %>% 
                  subset(y>2013) %>% 
                  select(y,
                         c
                         ) %>%
                  rename(year=y), 
                dat_p %>% 
                  select(year, 
                         price_buche, 
                         price_kg,
                         price_ton))

new_dat %>%
  mutate(year = as.factor(year))%>%
  ggplot(aes(x = c,
             y = price_ton 
             ))+
  geom_point(aes(color = year))+
  geom_smooth(method = "lm",
              formula = y ~ x, 
              se = T,
              color = 'black')+
  theme_bw()+
  scale_color_npg()+
  xlab('Catch of totoaba (metric tons)')+
  ylab('Price per ton of totoaba buche (in USD)')+
  labs(color = NULL)+
  theme(axis.text.x = element_text(angle = 70, 
                                   hjust = 1, 
                                   family = font_),
        axis.title.x = element_text(family = font_),
        axis.text.y = element_text(family = font_),
        axis.title.y = element_text(family = font_), 
        strip.text = element_text(family = font_),
        legend.position = 'bottom',
        legend.text = element_text(family = font_)
        )
#ggsave(here('visuals/supplementary_visuals', 'Figure10_price_data.jpg'))

```

## Extended data figure  : Von Bertlaffery Growth curves
```{r}
farmed_VBGF <- read.csv(here("data","inputs","aqua_par_farm.csv")) #Source: Semi-structured interviews
wild_VBGF <- read.csv(here("data","inputs","aqua_par_wild.csv")) #Source: INAPESCA 2020

farmed_VBGF = farmed_VBGF %>% 
  mutate(spec = 'Farmed')%>%
  rename(Target_Buche_Weight_g=Target_Buche_Weight_kg)
wild_VBGF = wild_VBGF %>% mutate(spec = 'Wild')
dat = rbind(farmed_VBGF, wild_VBGF)


buche_weight = 500

dat %>%
  ggplot(aes(y = Weight_Buche_g, x = Totoaba_Age, colour = spec))+
  geom_line(linewidth= 1.3)+
  geom_hline(yintercept = buche_weight, linetype = "dashed")+
  geom_segment(x =4.5, xend = 4.5, y=0, yend=buche_weight, colour='black', linetype="dashed" )+
  geom_segment(x =5,   xend = 5,   y=0, yend=buche_weight, colour='black', linetype="dashed" )+
  annotate("rect", xmin = 4.5, xmax = 5, ymin = 0, ymax = 500, alpha = .5) +
  #geom_segment(x = 8.3, xend=8.3,  y=0, yend=buche_weight, colour='black', linetype = "dashed") +
  scale_color_aaas()+
  theme_bw()+
  labs(colour = 'Growing conditions:')+
  ylab("Weight of swim bladder (in g)")+
  xlab('Age')+
  theme(axis.text.x = element_text(angle = 70, 
                                   hjust = 1, 
                                   family = font_),
        axis.title.x = element_text(family = font_),
        axis.text.y = element_text(family = font_),
        axis.title.y = element_text(family = font_), 
        strip.text = element_text(family = font_),
        legend.position = 'bottom',
        legend.text = element_text(family = font_),
        legend.title = element_text(family = font_)
        )
  
ggsave(here("visuals", 'sup_fig4_VBGF_growth.jpg'), width = 15, height = 22, units = 'cm')


```

# Illustration of the link between harvest and v

```{r, v func_}

a_m = alpha/beta
b_m = 1/beta

phi = (a_m - b_m*c)/(b_w*b_f*e)
eta = 2*W2*b_w*(2*b_w*b_f - e^2)
mu = (4*b_w*b_f - e^2)*sigma^2
theta = 2*W2*b_m
nu = 2*sigma^2
zeta = ((e*a_f + c*(e^2 - 2*b_w*b_f) + 2*b_f*a_w))/(b_f*e)
kappa = W1 * sigma * ((4*b_f*b_w - e^2)*(b_m - b_w)+e^2* b_w)
omega = 2*b_w*b_f*e*sigma^2
epsilon = 2*b_m*b_w*b_f*e*W2


v_func = function(x){
  v = phi*(eta + mu*x^2)/(theta + nu * x^2) - (kappa*x)/(omega*x^2 + epsilon) - zeta
  return(v)
}

component_one = function(x){
  to_ret = phi*(eta+mu*x^2)/(theta + nu*x^2)
  return(to_ret)
}

component_two = function(x){
  y = kappa*x/(omega*x^2 + epsilon)
  return(y)
}



K = calibration %>%
  subset(params == 'k')%>%
  select(value)%>%
  pull()

dat = data.frame(x = seq(0,K))%>%
  mutate(v_ = v_func(x))

dat_1 = dat %>%
  subset(x>=4000)

dat_rib1 = dat_1%>%
    subset(x<= 14470)
dat_rib2 = dat_1%>%
  subset(x>= 14470)

ggplot(dat_1, aes(x = x, y = v_))+
  geom_line()+
  #geom_vline(xintercept = 1/sigma*sqrt(W2*b_m))+
  geom_ribbon(data = dat_rib1, aes(ymin = v, ymax = v_, alpha = 0.2), fill = '#B22222', show.legend = F)+
  geom_ribbon(data = dat_rib2, aes(ymin = v, ymax = v_, alpha = 0.2), fill = '#248721', show.legend = F)+
  geom_hline(yintercept = v, linetype = 'dashed')+
  ylab('Threshold value of v e.g. v=F(x|c)')+
  xlab('Population')+
  theme(legend.position = 'bottom',
        axis.text.x = element_text(angle = 70, 
                                   hjust = 1, 
                                   family = font_),
        axis.title.x = element_text(family = font_),
        axis.text.y = element_text(family = font_),
        axis.title.y = element_text(family = font_), 
        strip.text = element_text(family = font_),
        legend.text = element_text(family = font_)
        )+
  theme_bw()
ggsave(here('visuals', 'supplementary_visuals', 'lemma2_bertrand.png'))
```
